src: remote: Introduce remote file
Working with Linux Kernel requires multiple deploys and works in a
remote machine. For this reason, this commit adds a new file named
`remote.sh` to keep multiple functions for handling the work with a
remote machine. Notice that this is part of a patchset that introduces
the deploy command. Finally, we have the unit test for this file.

Signed-off-by: Rodrigo Siqueira <rodrigosiqueiramelo@gmail.com>